refactor: add api versioning (v1)
