+++ src/helpers/container/index.ts
@@ -1,9 +1,9 @@
 import { container } from "tsyringe";
 
-import { ISensorRepositories } from "../../modules/Sensors/repositories/ISensorRepositories";
-import { SensorRepositoriesMongoDB } from "../../modules/Sensors/repositories/implementations/SensorRepositoriesMongoDB";
+import { ISensorRepositories } from "../../modules/v1/Sensors/repositories/ISensorRepositories";
+import { SensorRepositoriesMongoDB } from "../../modules/v1/Sensors/repositories/implementations/SensorRepositoriesMongoDB";
 
 container.registerSingleton<ISensorRepositories>(
   "SensorRepositories",
   SensorRepositoriesMongoDB
-);  
+);
