+++ src/modules/v1/Sensors/entities/Sensor.ts
