+++ src/modules/v1/Sensors/schemas/SensorSchema.ts
