+++ src/modules/v1/Sensors/useCases/delete/deleteSensor.controller.ts
