+++ src/modules/v1/Sensors/useCases/findById/findById.controller.ts
